fastsite/pyproject.toml
[project]
name = "fastsite"
version = "0.1.0"
description = "My FastAPI Web App"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "icecream>=2.1.5",
    "passlib>=1.7.4",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]

[tool.setuptools]
packages = ["src"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "hypothesis>=6.138.2",
    "pytest>=8.4.1",
    "pytest-dotenv>=0.5.2",
    "pytest-mock>=3.14.1",
    "schemathesis>=4.1.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-ra -q"
env_files = [".env"]

